GPU Analysis
inPolygin, getClosestPoint関数を実装
CoPの計算が正しくできることを確認

+++ capturability/include/vector.h
@@ -15,17 +15,17 @@ struct Vector2 {
 
   void clear();
 
-  void setPolar(float radius, float theta);
-  void setCartesian(float x, float y);
+  void setPolar(double radius, double theta);
+  void setCartesian(double x, double y);
 
   void printPolar();
   void printPolar(std::string str);
   void printCartesian();
   void printCartesian(std::string str);
 
-  float r, th;
-  float x, y;
-  float norm();
+  double r, th;
+  double x, y;
+  double norm();
 
   Vector2 normal();
 
@@ -33,10 +33,10 @@ struct Vector2 {
 
   Vector2 operator+(const Vector2 &v);
   Vector2 operator-(const Vector2 &v) const;
-  float operator  %(const Vector2 &v);
-  Vector2 operator*(const float &d);
-  float operator  *(const Vector2 &v);
-  Vector2 operator/(const float &d);
+  double operator %(const Vector2 &v);
+  Vector2 operator*(const double &d);
+  double operator *(const Vector2 &v);
+  Vector2 operator/(const double &d);
 
 private:
   void cartesianToPolar();
@@ -47,7 +47,7 @@ private:
 typedef Vector2 vec2_t;
 typedef std::vector<Vector2> arr2_t;
 
-Vector2 operator*(const float &d, const Vector2 &v);
+Vector2 operator*(const double &d, const Vector2 &v);
 
 } // namespace Capt
